Fix [ 685603 ] warning: cast to pointer from integer of different size and [ 654048 ] patch: incorrect cast from/to pointers.
Found a couple more items with the 64-bit compiler.  Tests OK.

+++ tidy/src/streamio.c
@@ -70,14 +70,14 @@ static StreamOut stdoutStreamOut =
 StreamOut* StdErrOutput()
 {
   if ( stderrStreamOut.sink.sinkData == 0 )
-      stderrStreamOut.sink.sinkData = (uint) stderr;
+      stderrStreamOut.sink.sinkData = (ulong) stderr;
   return &stderrStreamOut;
 }
 
 StreamOut* StdOutOutput()
 {
   if ( stdoutStreamOut.sink.sinkData == 0 )
-      stdoutStreamOut.sink.sinkData = (uint) stdout;
+      stdoutStreamOut.sink.sinkData = (ulong) stdout;
   return &stdoutStreamOut;
 }
 
@@ -822,7 +822,7 @@ Bool tidyInitSource( TidyInputSource*  source,
   Bool ok = ( source && srcData && gbFunc && ugbFunc && endFunc );
   if ( ok )
   {
-    source->sourceData = (uint) srcData;
+    source->sourceData = (ulong) srcData;
     source->getByte    = gbFunc;
     source->ungetByte  = ugbFunc;
     source->eof        = endFunc;
@@ -837,7 +837,7 @@ Bool tidyInitSink( TidyOutputSink* sink,
   Bool ok = ( sink && snkData && pbFunc );
   if ( ok )
   {
-    sink->sinkData = (uint) snkData;
+    sink->sinkData = (ulong) snkData;
     sink->putByte  = pbFunc;
   }
   return ok;
